[ASTGen] Move macros related code to a new dedicated module
Introduce 'swiftMacros' module.
