+++ lib/ASTGen/Package.swift
@@ -54,7 +54,6 @@ let package = Package(
     .target(
       name: "swiftASTGen",
       dependencies: [
-        .product(name: "_SwiftCompilerPluginMessageHandling", package: "swift-syntax"),
         .product(name: "SwiftDiagnostics", package: "swift-syntax"),
         .product(name: "SwiftIfConfig", package: "swift-syntax"),
         .product(name: "SwiftOperators", package: "swift-syntax"),
@@ -66,6 +65,20 @@ let package = Package(
         "_CompilerRegexParser",
       ],
       path: "Sources/ASTGen",
+      swiftSettings: swiftSetttings
+    ),
+    .target(
+      name: "swiftMacros",
+      dependencies: [
+        "swiftASTGen",
+        .product(name: "_SwiftCompilerPluginMessageHandling", package: "swift-syntax"),
+        .product(name: "SwiftSyntax", package: "swift-syntax"),
+        .product(name: "SwiftDiagnostics", package: "swift-syntax"),
+        .product(name: "SwiftParser", package: "swift-syntax"),
+        .product(name: "SwiftOperators", package: "swift-syntax"),
+        .product(name: "SwiftSyntaxMacroExpansion", package: "swift-syntax"),
+      ],
+      path: "Sources/Macros",
       swiftSettings: swiftSetttings
     ),
     .target(
