+++ lib/ASTGen/Sources/ASTGen/SourceFile.swift
@@ -25,10 +25,10 @@ public struct ExportedSourceFile {
   public let buffer: UnsafeBufferPointer<UInt8>
 
   /// The name of the enclosing module.
-  let moduleName: String
+  public let moduleName: String
 
   /// The name of the source file being parsed.
-  let fileName: String
+  public let fileName: String
 
   /// The syntax tree for the complete source file.
   public let syntax: SourceFileSyntax
@@ -36,7 +36,7 @@ public struct ExportedSourceFile {
   /// A source location converter to convert `AbsolutePosition`s in `syntax` to line/column locations.
   ///
   /// Cached so we don't need to re-build the line table every time we need to convert a position.
-  let sourceLocationConverter: SourceLocationConverter
+  public let sourceLocationConverter: SourceLocationConverter
 
   /// Configured regions for this source file.
   ///
@@ -199,4 +199,65 @@ public func emitParserDiagnostics(
 
     return anyDiags ? 1 : 0
   }
+}
+
+/// Retrieve a syntax node in the given source file, with the given type.
+public func findSyntaxNodeInSourceFile<Node: SyntaxProtocol>(
+  sourceFilePtr: UnsafeRawPointer,
+  sourceLocationPtr: UnsafePointer<UInt8>?,
+  type: Node.Type,
+  wantOutermost: Bool = false
+) -> Node? {
+  guard let sourceLocationPtr = sourceLocationPtr else {
+    return nil
+  }
+
+  let sourceFilePtr = sourceFilePtr.assumingMemoryBound(to: ExportedSourceFile.self)
+
+  // Find the offset.
+  let buffer = sourceFilePtr.pointee.buffer
+  let offset = sourceLocationPtr - buffer.baseAddress!
+  if offset < 0 || offset >= buffer.count {
+    print("source location isn't inside this buffer")
+    return nil
+  }
+
+  // Find the token at that offset.
+  let sf = sourceFilePtr.pointee.syntax
+  guard let token = sf.token(at: AbsolutePosition(utf8Offset: offset)) else {
+    print("couldn't find token at offset \(offset)")
+    return nil
+  }
+
+  var currentSyntax = Syntax(token)
+  var resultSyntax: Node? = nil
+  while let parentSyntax = currentSyntax.parent {
+    currentSyntax = parentSyntax
+    if let typedParent = currentSyntax.as(type) {
+      resultSyntax = typedParent
+      break
+    }
+  }
+
+  // If we didn't find anything, complain and fail.
+  guard var resultSyntax else {
+    print("unable to find node: \(token.debugDescription)")
+    return nil
+  }
+
+  // If we want the outermost node, keep looking.
+  // E.g. for 'foo.bar' we want the member ref expression instead of the
+  // identifier expression.
+  if wantOutermost {
+    while let parentSyntax = currentSyntax.parent,
+      parentSyntax.position == resultSyntax.position
+    {
+      currentSyntax = parentSyntax
+      if let typedParent = currentSyntax.as(type) {
+        resultSyntax = typedParent
+      }
+    }
+  }
+
+  return resultSyntax
 }
